requirements.txt
Django==1.7c2
Pillow==2.5.1
Werkzeug==0.9.6
django-crispy-forms==1.4.0
django-extensions==1.3.9
six==1.7.3
